pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["extractors*", "normalizers*", "domain*", "storage*", "mcp_server*", "scripts*"]

[project]
name = "ia-invest"
version = "0.1.0"
description = "Local-first investment portfolio ingestion, consolidation and AI-agent analysis"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "pdfplumber>=0.11",
    "pandas>=2.2",
    "mcp>=1.0",
    "structlog>=24.0",
    "python-dateutil>=2.9",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "types-python-dateutil>=2.9",
]

[project.scripts]
ia-invest-import = "scripts.import_all:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--tb=short -q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
